SingleCountryDetailsPage cleanup
